Silence compiler warnings about comparing signed and unsigned integers.
git-svn-id: https://opensg.vrsource.org/svn/trunk@701 4683daeb-ad0f-0410-a623-93161e962ae5

+++ Source/System/FieldContainer/Mixins/OSGDataSlotMixin.inl
@@ -44,7 +44,7 @@ ValuePtr DataSlotMixin<ParentT>::getData(Int32 iSlotId)
 {
     ValuePtr returnValue = NULL;
 
-    if(iSlotId >= 0 && iSlotId < _mfData.size())
+    if(iSlotId >= 0 && static_cast<UInt32>(iSlotId) < _mfData.size())
     {
         FieldBundleP pTmp = _mfData[iSlotId];
 
@@ -60,7 +60,7 @@ void DataSlotMixin<ParentT>::setDataX(FieldBundleP pBundle, Int32 iSlotId)
     if(iSlotId < 0)
         return;
 
-    if(iSlotId <= _mfData.size())
+    if(static_cast<UInt32>(iSlotId) <= _mfData.size())
     {
         _mfData.resize(iSlotId + 1, NULL);
     }
